backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
anthropic==0.34.2
redis==5.0.8
python-multipart==0.0.12
Pillow==10.4.0
